requirements.txt
numpy>=1.23
pandas>=1.5
pyarrow>=11.0
scikit-learn>=1.2
plotly>=5.14
google-generativeai>=0.5
